提炼 静/动态彩带特效、星空雨特效 到public，优化 线段粒子特效和樱花特效 在public中的代码 特效已经全部封装到public当中
