+++ public/js/nest.js
@@ -5,9 +5,9 @@
  * @param config
  */
 
-const id = '__next'
+const idNest = '__nest'
 function createNest() {
-  const e = document.getElementById('__next')
+  const e = document.getElementById(idNest)
   if(!e) return
   function n(e, n, t) {
     return e.getAttribute(n) || t
@@ -108,7 +108,7 @@ function createNest() {
 }
 
 function destroyNest() {
-  const nest = document.getElementById(id)
+  const nest = document.getElementById(idNest)
   if(nest && nest.parentNode){
     nest.parentNode.removeChild(nest)
   }
